requirements.txt
celery==5.2.7
celery[sqs]==5.2.7
django==4.1.5
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-celery-results==2.4.0
django-cors-headers==3.13.0
django-redis==5.2.0
factory-boy==3.2.1
Faker==16.6.1
gunicorn==20.1.0
psycopg2-binary==2.9.5
pytz==2022.7
python-dotenv==0.21.0
redis==4.4.2
requests==2.28.2
watchdog==2.2.1
